+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q002/Solution2.java
@@ -1,4 +1,6 @@
-package com.noob.algorithm.leetcode.q02;
+package com.noob.algorithm.leetcode.q002;
+
+import com.noob.algorithm.base.dataStructure.tree.ListNode;
 
 /**
  * 2-两数相加（超出内存限制）
